requirements.txt
yt-dlp>=2024.1.0
opencv-python>=4.8.0
Pillow>=10.0.0
ultralytics>=8.0.0
imagehash>=4.3.0
boto3>=1.34.0
